+++ bindings/c/manifoldc.cpp
@@ -35,25 +35,27 @@
 using namespace manifold;
 
 namespace {
-ManifoldMeshGL *level_set(void *mem, float (*sdf)(float, float, float),
-                          ManifoldBox *bounds, float edge_length, float level,
+ManifoldMeshGL *level_set(void *mem, double (*sdf)(double, double, double),
+                          ManifoldBox *bounds, double edge_length, double level,
                           bool seq) {
   // typing with std::function rather than auto compiles when CUDA is on,
   // passing it into GPU (and crashing) is avoided dynamically in `sdf.h`
-  std::function<float(glm::vec3)> fun = [sdf](glm::vec3 v) {
+  std::function<double(glm::dvec3)> fun = [sdf](glm::dvec3 v) {
     return (sdf(v.x, v.y, v.z));
   };
   auto mesh = LevelSet(fun, *from_c(bounds), edge_length, level, !seq);
   return to_c(new (mem) MeshGL(mesh));
 }
-ManifoldMeshGL *level_set_context(
-    void *mem, float (*sdf_context)(float, float, float, void *),
-    ManifoldBox *bounds, float edge_length, float level, bool seq, void *ctx) {
+ManifoldMeshGL *level_set_context(void *mem,
+                                  double (*sdf_context)(double, double, double,
+                                                        void *),
+                                  ManifoldBox *bounds, double edge_length,
+                                  double level, bool seq, void *ctx) {
   // Bind function with context argument to one without
   using namespace std::placeholders;
-  std::function<float(float, float, float)> sdf =
+  std::function<double(double, double, double)> sdf =
       std::bind(sdf_context, _1, _2, _3, ctx);
-  std::function<float(glm::vec3)> fun = [sdf](glm::vec3 v) {
+  std::function<double(glm::dvec3)> fun = [sdf](glm::dvec3 v) {
     return (sdf(v.x, v.y, v.z));
   };
   auto mesh = LevelSet(fun, *from_c(bounds), edge_length, level, !seq);
@@ -67,7 +69,7 @@ extern "C" {
 
 ManifoldSimplePolygon *manifold_simple_polygon(void *mem, ManifoldVec2 *ps,
                                                size_t length) {
-  auto vec = new (mem) std::vector<glm::vec2>;
+  auto vec = new (mem) std::vector<glm::dvec2>;
   for (int i = 0; i < length; ++i) {
     vec->push_back({ps[i].x, ps[i].y});
   }
@@ -185,9 +187,9 @@ ManifoldManifoldPair manifold_split(void *mem_first, void *mem_second,
 
 ManifoldManifoldPair manifold_split_by_plane(void *mem_first, void *mem_second,
                                              ManifoldManifold *m,
-                                             float normal_x, float normal_y,
-                                             float normal_z, float offset) {
-  auto normal = glm::vec3(normal_x, normal_y, normal_z);
+                                             double normal_x, double normal_y,
+                                             double normal_z, double offset) {
+  auto normal = glm::dvec3(normal_x, normal_y, normal_z);
   auto pair = from_c(m)->SplitByPlane(normal, offset);
   auto first = new (mem_first) Manifold(pair.first);
   auto second = new (mem_second) Manifold(pair.second);
@@ -195,15 +197,15 @@ ManifoldManifoldPair manifold_split_by_plane(void *mem_first, void *mem_second,
 }
 
 ManifoldManifold *manifold_trim_by_plane(void *mem, ManifoldManifold *m,
-                                         float normal_x, float normal_y,
-                                         float normal_z, float offset) {
-  auto normal = glm::vec3(normal_x, normal_y, normal_z);
+                                         double normal_x, double normal_y,
+                                         double normal_z, double offset) {
+  auto normal = glm::dvec3(normal_x, normal_y, normal_z);
   auto trimmed = from_c(m)->TrimByPlane(normal, offset);
   return to_c(new (mem) Manifold(trimmed));
 }
 
 ManifoldCrossSection *manifold_slice(void *mem, ManifoldManifold *m,
-                                     float height) {
+                                     double height) {
   auto poly = from_c(m)->Slice(height);
   return to_c(new (mem) CrossSection(poly));
 }
@@ -225,7 +227,7 @@ ManifoldManifold *manifold_batch_hull(void *mem, ManifoldManifoldVec *ms) {
 
 ManifoldManifold *manifold_hull_pts(void *mem, ManifoldVec3 *ps,
                                     size_t length) {
-  std::vector<glm::vec3> vec(length);
+  std::vector<glm::dvec3> vec(length);
   for (int i = 0; i < length; ++i) {
     vec[i] = {ps[i].x, ps[i].y, ps[i].z};
   }
@@ -233,72 +235,73 @@ ManifoldManifold *manifold_hull_pts(void *mem, ManifoldVec3 *ps,
   return to_c(new (mem) Manifold(hulled));
 }
 
-ManifoldManifold *manifold_translate(void *mem, ManifoldManifold *m, float x,
-                                     float y, float z) {
-  auto v = glm::vec3(x, y, z);
+ManifoldManifold *manifold_translate(void *mem, ManifoldManifold *m, double x,
+                                     double y, double z) {
+  auto v = glm::dvec3(x, y, z);
   auto translated = from_c(m)->Translate(v);
   return to_c(new (mem) Manifold(translated));
 }
 
-ManifoldManifold *manifold_rotate(void *mem, ManifoldManifold *m, float x,
-                                  float y, float z) {
+ManifoldManifold *manifold_rotate(void *mem, ManifoldManifold *m, double x,
+                                  double y, double z) {
   auto rotated = from_c(m)->Rotate(x, y, z);
   return to_c(new (mem) Manifold(rotated));
 }
 
-ManifoldManifold *manifold_scale(void *mem, ManifoldManifold *m, float x,
-                                 float y, float z) {
-  auto s = glm::vec3(x, y, z);
+ManifoldManifold *manifold_scale(void *mem, ManifoldManifold *m, double x,
+                                 double y, double z) {
+  auto s = glm::dvec3(x, y, z);
   auto scaled = from_c(m)->Scale(s);
   return to_c(new (mem) Manifold(scaled));
 }
 
-ManifoldManifold *manifold_transform(void *mem, ManifoldManifold *m, float x1,
-                                     float y1, float z1, float x2, float y2,
-                                     float z2, float x3, float y3, float z3,
-                                     float x4, float y4, float z4) {
-  auto mat = glm::mat4x3(x1, y1, z1, x2, y2, z2, x3, y3, z3, x4, y4, z4);
+ManifoldManifold *manifold_transform(void *mem, ManifoldManifold *m, double x1,
+                                     double y1, double z1, double x2, double y2,
+                                     double z2, double x3, double y3, double z3,
+                                     double x4, double y4, double z4) {
+  auto mat = glm::dmat4x3(x1, y1, z1, x2, y2, z2, x3, y3, z3, x4, y4, z4);
   auto transformed = from_c(m)->Transform(mat);
   return to_c(new (mem) Manifold(transformed));
 }
 
-ManifoldManifold *manifold_mirror(void *mem, ManifoldManifold *m, float nx,
-                                  float ny, float nz) {
+ManifoldManifold *manifold_mirror(void *mem, ManifoldManifold *m, double nx,
+                                  double ny, double nz) {
   auto mirrored = from_c(m)->Mirror({nx, ny, nz});
   return to_c(new (mem) Manifold(mirrored));
 }
 
 ManifoldManifold *manifold_warp(void *mem, ManifoldManifold *m,
-                                ManifoldVec3 (*fun)(float, float, float)) {
-  std::function<void(glm::vec3 & v)> warp = [fun](glm::vec3 &v) {
+                                ManifoldVec3 (*fun)(double, double, double)) {
+  std::function<void(glm::dvec3 & v)> warp = [fun](glm::dvec3 &v) {
     v = from_c(fun(v.x, v.y, v.z));
   };
   auto warped = from_c(m)->Warp(warp);
   return to_c(new (mem) Manifold(warped));
 }
 
-ManifoldMeshGL *manifold_level_set(void *mem, float (*sdf)(float, float, float),
-                                   ManifoldBox *bounds, float edge_length,
-                                   float level) {
+ManifoldMeshGL *manifold_level_set(void *mem,
+                                   double (*sdf)(double, double, double),
+                                   ManifoldBox *bounds, double edge_length,
+                                   double level) {
   return level_set(mem, sdf, bounds, edge_length, level, false);
 }
 
 ManifoldMeshGL *manifold_level_set_seq(void *mem,
-                                       float (*sdf)(float, float, float),
-                                       ManifoldBox *bounds, float edge_length,
-                                       float level) {
+                                       double (*sdf)(double, double, double),
+                                       ManifoldBox *bounds, double edge_length,
+                                       double level) {
   return level_set(mem, sdf, bounds, edge_length, level, true);
 }
 
 ManifoldMeshGL *manifold_level_set_context(
-    void *mem, float (*sdf)(float, float, float, void *), ManifoldBox *bounds,
-    float edge_length, float level, void *ctx) {
+    void *mem, double (*sdf)(double, double, double, void *),
+    ManifoldBox *bounds, double edge_length, double level, void *ctx) {
   return level_set_context(mem, sdf, bounds, edge_length, level, false, ctx);
 }
 
 ManifoldMeshGL *manifold_level_set_seq_context(
-    void *mem, float (*sdf)(float, float, float, void *), ManifoldBox *bounds,
-    float edge_length, float level, void *ctx) {
+    void *mem, double (*sdf)(double, double, double, void *),
+    ManifoldBox *bounds, double edge_length, double level, void *ctx) {
   return level_set_context(mem, sdf, bounds, edge_length, level, true, ctx);
 }
 
@@ -320,28 +323,28 @@ ManifoldManifold *manifold_tetrahedron(void *mem) {
   return to_c(new (mem) Manifold(m));
 }
 
-ManifoldManifold *manifold_cube(void *mem, float x, float y, float z,
+ManifoldManifold *manifold_cube(void *mem, double x, double y, double z,
                                 int center) {
-  auto size = glm::vec3(x, y, z);
+  auto size = glm::dvec3(x, y, z);
   auto m = Manifold::Cube(size, center);
   return to_c(new (mem) Manifold(m));
 }
 
-ManifoldManifold *manifold_cylinder(void *mem, float height, float radius_low,
-                                    float radius_high, int circular_segments,
+ManifoldManifold *manifold_cylinder(void *mem, double height, double radius_low,
+                                    double radius_high, int circular_segments,
                                     int center) {
   auto m = Manifold::Cylinder(height, radius_low, radius_high,
                               circular_segments, center);
   return to_c(new (mem) Manifold(m));
 }
 
-ManifoldManifold *manifold_sphere(void *mem, float radius,
+ManifoldManifold *manifold_sphere(void *mem, double radius,
                                   int circular_segments) {
   auto m = Manifold::Sphere(radius, circular_segments);
   return to_c(new (mem) Manifold(m));
 }
 
-ManifoldMeshGL *manifold_meshgl(void *mem, float *vert_props, size_t n_verts,
+ManifoldMeshGL *manifold_meshgl(void *mem, double *vert_props, size_t n_verts,
                                 size_t n_props, uint32_t *tri_verts,
                                 size_t n_tris) {
   auto mesh = new (mem) MeshGL();
@@ -351,10 +354,10 @@ ManifoldMeshGL *manifold_meshgl(void *mem, float *vert_props, size_t n_verts,
   return to_c(mesh);
 }
 
-ManifoldMeshGL *manifold_meshgl_w_tangents(void *mem, float *vert_props,
+ManifoldMeshGL *manifold_meshgl_w_tangents(void *mem, double *vert_props,
                                            size_t n_verts, size_t n_props,
                                            uint32_t *tri_verts, size_t n_tris,
-                                           float *halfedge_tangent) {
+                                           double *halfedge_tangent) {
   auto mesh = new (mem) MeshGL();
   mesh->numProp = n_props;
   mesh->vertProperties = vector_of_array(vert_props, n_verts * n_props);
@@ -364,7 +367,7 @@ ManifoldMeshGL *manifold_meshgl_w_tangents(void *mem, float *vert_props,
 }
 
 ManifoldManifold *manifold_smooth(void *mem, ManifoldMeshGL *mesh,
-                                  int *half_edges, float *smoothness,
+                                  int *half_edges, double *smoothness,
                                   int n_edges) {
   auto smooth = std::vector<Smoothness>();
   for (int i = 0; i < n_edges; ++i) {
@@ -380,10 +383,10 @@ ManifoldManifold *manifold_of_meshgl(void *mem, ManifoldMeshGL *mesh) {
 }
 
 ManifoldManifold *manifold_extrude(void *mem, ManifoldCrossSection *cs,
-                                   float height, int slices,
-                                   float twist_degrees, float scale_x,
-                                   float scale_y) {
-  auto scale = glm::vec2(scale_x, scale_y);
+                                   double height, int slices,
+                                   double twist_degrees, double scale_x,
+                                   double scale_y) {
+  auto scale = glm::dvec2(scale_x, scale_y);
   auto m = Manifold::Extrude(*from_c(cs), height, slices, twist_degrees, scale);
   return to_c(new (mem) Manifold(m));
 }
@@ -450,7 +453,7 @@ size_t manifold_meshgl_tangent_length(ManifoldMeshGL *m) {
   return from_c(m)->halfedgeTangent.size();
 }
 
-float *manifold_meshgl_vert_properties(void *mem, ManifoldMeshGL *m) {
+double *manifold_meshgl_vert_properties(void *mem, ManifoldMeshGL *m) {
   return copy_data(mem, from_c(m)->vertProperties);
 }
 
@@ -474,7 +477,7 @@ uint32_t *manifold_meshgl_run_original_id(void *mem, ManifoldMeshGL *m) {
   return copy_data(mem, from_c(m)->runOriginalID);
 }
 
-float *manifold_meshgl_run_transform(void *mem, ManifoldMeshGL *m) {
+double *manifold_meshgl_run_transform(void *mem, ManifoldMeshGL *m) {
   return copy_data(mem, from_c(m)->runTransform);
 }
 
@@ -482,7 +485,7 @@ uint32_t *manifold_meshgl_face_id(void *mem, ManifoldMeshGL *m) {
   return copy_data(mem, from_c(m)->faceID);
 }
 
-float *manifold_meshgl_halfedge_tangent(void *mem, ManifoldMeshGL *m) {
+double *manifold_meshgl_halfedge_tangent(void *mem, ManifoldMeshGL *m) {
   return copy_data(mem, from_c(m)->halfedgeTangent);
 }
 
@@ -516,17 +519,19 @@ ManifoldBox *manifold_bounding_box(void *mem, ManifoldManifold *m) {
   return to_c(new (mem) Box(box));
 }
 
-float manifold_precision(ManifoldManifold *m) { return from_c(m)->Precision(); }
+double manifold_precision(ManifoldManifold *m) {
+  return from_c(m)->Precision();
+}
 
 uint32_t manifold_reserve_ids(uint32_t n) { return Manifold::ReserveIDs(n); }
 
 ManifoldManifold *manifold_set_properties(void *mem, ManifoldManifold *m,
                                           int num_prop,
-                                          void (*fun)(float *new_prop,
+                                          void (*fun)(double *new_prop,
                                                       ManifoldVec3 position,
-                                                      const float *old_prop)) {
-  std::function<void(float *, glm::vec3, const float *)> f =
-      [fun](float *new_prop, glm::vec3 v, const float *old_prop) {
+                                                      const double *old_prop)) {
+  std::function<void(double *, glm::dvec3, const double *)> f =
+      [fun](double *new_prop, glm::dvec3 v, const double *old_prop) {
         fun(new_prop, to_c(v), old_prop);
       };
   auto man = from_c(m)->SetProperties(num_prop, f);
@@ -541,11 +546,11 @@ ManifoldManifold *manifold_calculate_curvature(void *mem, ManifoldManifold *m,
 
 // Static Quality Globals
 
-void manifold_set_min_circular_angle(float degrees) {
+void manifold_set_min_circular_angle(double degrees) {
   Quality::SetMinCircularAngle(degrees);
 }
 
-void manifold_set_min_circular_edge_length(float length) {
+void manifold_set_min_circular_edge_length(double length) {
   Quality::SetMinCircularEdgeLength(length);
 }
 
@@ -553,7 +558,7 @@ void manifold_set_circular_segments(int number) {
   Quality::SetCircularSegments(number);
 }
 
-int manifold_get_circular_segments(float radius) {
+int manifold_get_circular_segments(double radius) {
   return Quality::GetCircularSegments(radius);
 }
