+++ meshIO/src/meshIO.cpp
@@ -45,7 +45,7 @@ aiScene* CreateScene(const ExportOptions& options) {
   aiMaterial* material = scene->mMaterials[0];
   material->AddProperty(&options.mat.roughness, 1, AI_MATKEY_ROUGHNESS_FACTOR);
   material->AddProperty(&options.mat.metalness, 1, AI_MATKEY_METALLIC_FACTOR);
-  const glm::vec4& color = options.mat.color;
+  const glm::dvec4& color = options.mat.color;
   aiColor4D baseColor(color.r, color.g, color.b, color.a);
   material->AddProperty(&baseColor, 1, AI_MATKEY_COLOR_DIFFUSE);
 
@@ -148,8 +148,8 @@ Mesh ImportMesh(const std::string& filename, bool forceCleanup) {
     const aiMesh* mesh_i = scene->mMeshes[i];
     for (int j = 0; j < mesh_i->mNumVertices; ++j) {
       const aiVector3D vert = mesh_i->mVertices[j];
-      mesh_out.vertPos.push_back(isYup ? glm::vec3(vert.z, vert.x, vert.y)
-                                       : glm::vec3(vert.x, vert.y, vert.z));
+      mesh_out.vertPos.push_back(isYup ? glm::dvec3(vert.z, vert.x, vert.y)
+                                       : glm::dvec3(vert.x, vert.y, vert.z));
     }
     for (int j = 0; j < mesh_i->mNumFaces; ++j) {
       const aiFace face = mesh_i->mFaces[j];
@@ -215,12 +215,12 @@ void ExportMesh(const std::string& filename, const MeshGL& mesh,
   if (hasColor) mesh_out->mColors[0] = new aiColor4D[mesh_out->mNumVertices];
 
   for (int i = 0; i < mesh_out->mNumVertices; ++i) {
-    glm::vec3 v;
+    glm::dvec3 v;
     for (int j : {0, 1, 2}) v[j] = mesh.vertProperties[i * mesh.numProp + j];
     mesh_out->mVertices[i] =
         isYup ? aiVector3D(v.y, v.z, v.x) : aiVector3D(v.x, v.y, v.z);
     if (!options.faceted) {
-      glm::vec3 n;
+      glm::dvec3 n;
       for (int j : {0, 1, 2})
         n[j] = mesh.vertProperties[i * mesh.numProp +
                                    options.mat.normalChannels[j]];
@@ -228,7 +228,7 @@ void ExportMesh(const std::string& filename, const MeshGL& mesh,
           isYup ? aiVector3D(n.y, n.z, n.x) : aiVector3D(n.x, n.y, n.z);
     }
     if (hasColor) {
-      glm::vec4 c;
+      glm::dvec4 c;
       for (int j : {0, 1, 2, 3})
         c[j] = options.mat.colorChannels[j] < 0
                    ? 1
@@ -296,16 +296,16 @@ void ExportMesh(const std::string& filename, const Mesh& mesh,
   }
 
   for (int i = 0; i < mesh_out->mNumVertices; ++i) {
-    const glm::vec3& v = mesh.vertPos[i];
+    const glm::dvec3& v = mesh.vertPos[i];
     mesh_out->mVertices[i] =
         isYup ? aiVector3D(v.y, v.z, v.x) : aiVector3D(v.x, v.y, v.z);
     if (!options.faceted) {
-      const glm::vec3& n = mesh.vertNormal[i];
+      const glm::dvec3& n = mesh.vertNormal[i];
       mesh_out->mNormals[i] =
           isYup ? aiVector3D(n.y, n.z, n.x) : aiVector3D(n.x, n.y, n.z);
     }
     if (!options.mat.vertColor.empty()) {
-      const glm::vec4& c = options.mat.vertColor[i];
+      const glm::dvec4& c = options.mat.vertColor[i];
       mesh_out->mColors[0][i] = aiColor4D(c.r, c.g, c.b, c.a);
     }
   }
